update vite-react-ts example with new pkg features
