+++ examples/vite-react-ts/tailwind.config.ts
@@ -2,7 +2,7 @@ import type { Config } from "tailwindcss";
 import { themeColors } from "./src/themes/theme-colors";
 
 export default {
-    content: ["./index.html", ".\\src\\**\\*.{js,ts,jsx,tsx}"],
+    content: ["./index.html", "./src\\**\\*.{js,ts,jsx,tsx}"],
     darkMode: "class",
     theme: {
         extend: {
