+++ examples/vite-react-ts/tsconfig.app.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/themes/hue-palettes.ts","./src/themes/theme-colors.ts"],"version":"5.6.3"}
