+++ examples/vite-react-ts/tsconfig.node.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["./vite.config.ts"],"version":"5.6.3"}
